+++ src/main/java/pl/asie/computronics/oc/RobotUpgradeCamera.java
@@ -28,7 +28,9 @@ public class RobotUpgradeCamera extends ManagedEnvironment {
 		return l;
 	}
 	
-    @Callback(direct = true, limit = CALL_LIMIT)
+    @Callback(doc = "function([x:number, y:number]):number; "
+			+ "Returns the distance to the block the ray is shot at with the specified x-y offset, "
+			+ "or of the block directly in front", direct = true, limit = CALL_LIMIT)
     public Object[] distance(Context context, Arguments args) {
     	float x = 0.0f;
     	float y = 0.0f;
@@ -41,7 +43,9 @@ public class RobotUpgradeCamera extends ManagedEnvironment {
 		return new Object[]{camera.getDistance()};
     }
     
-    @Callback(direct = true, limit = CALL_LIMIT)
+    @Callback(doc = "function([x:number, y:number]):number; "
+		+ "Returns the distance to the block the ray is shot at with the specified x-y offset facing upwards, "
+		+ "or of the block directly above", direct = true, limit = CALL_LIMIT)
     public Object[] distanceUp(Context context, Arguments args) {
     	float x = 0.0f;
     	float y = 0.0f;
@@ -54,7 +58,9 @@ public class RobotUpgradeCamera extends ManagedEnvironment {
 		return new Object[]{camera.getDistance()};
     }
     
-    @Callback(direct = true, limit = CALL_LIMIT)
+    @Callback(doc = "function([x:number, y:number]):number; "
+		+ "Returns the distance to the block the ray is shot at with the specified x-y offset facing downwards, "
+		+ "or of the block directly below", direct = true, limit = CALL_LIMIT)
     public Object[] distanceDown(Context context, Arguments args) {
     	float x = 0.0f;
     	float y = 0.0f;