+++ src/main/java/pl/asie/computronics/tile/TileLocomotiveRelay.java
@@ -1,5 +1,6 @@
 package pl.asie.computronics.tile;
 
+import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.Optional;
 import dan200.computercraft.api.lua.ILuaContext;
 import dan200.computercraft.api.lua.LuaException;
@@ -7,21 +8,18 @@ import dan200.computercraft.api.peripheral.IComputerAccess;
 import li.cil.oc.api.machine.Arguments;
 import li.cil.oc.api.machine.Callback;
 import li.cil.oc.api.machine.Context;
-import mods.railcraft.api.carts.CartTools;
+import li.cil.oc.api.network.Connector;
 import mods.railcraft.common.carts.EntityLocomotiveElectric;
 import mods.railcraft.common.items.ItemTicket;
 import mods.railcraft.common.items.ItemTicketGold;
 import mods.railcraft.common.util.misc.MiscTools;
-import net.minecraft.entity.item.EntityMinecart;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.util.AxisAlignedBB;
-import pl.asie.computronics.Computronics;
+import pl.asie.computronics.integration.railcraft.LocomotiveManager;
 import pl.asie.computronics.reference.Config;
 import pl.asie.computronics.reference.Mods;
 
-import java.util.List;
-import java.util.Locale;
+import java.lang.ref.WeakReference;
 import java.util.UUID;
 
 /**
@@ -29,120 +27,89 @@ import java.util.UUID;
  */
 public class TileLocomotiveRelay extends TileEntityPeripheralBase {
 
-	private EntityLocomotiveElectric locomotive;
-	private double locomotiveX, locomotiveY, locomotiveZ;
-	private boolean isInitialized = false, isBound = false;
+	private WeakReference<EntityLocomotiveElectric> locomotive;
+	//private boolean isInitialized = false;
+	private boolean isBound = false;
+
 	private UUID uuid;
 
 	public TileLocomotiveRelay() {
-		super("locomotive_relay");
+		super("locomotive_relay", Config.LOCOMOTIVE_RELAY_BASE_POWER * 10);
 	}
 
 	public void setLocomotive(EntityLocomotiveElectric loco) {
-		this.locomotive = loco;
-		this.locomotiveX = this.locomotive.posX;
-		this.locomotiveY = this.locomotive.posY;
-		this.locomotiveZ = this.locomotive.posZ;
+		this.locomotive = new WeakReference<EntityLocomotiveElectric>(loco);
 		this.isBound = true;
-		this.uuid = this.locomotive.getUniqueID();
+		this.uuid = loco.getPersistentID();
 	}
 
 	public EntityLocomotiveElectric getLocomotive() {
-		return this.locomotive;
+		return this.locomotive != null ? this.locomotive.get() : null;
 	}
 
 	public boolean isBound() {
 		return this.isBound;
 	}
 
+	public boolean unbind() {
+		if(this.isBound) {
+			this.isBound = false;
+			this.locomotive = null;
+			this.uuid = null;
+			return true;
+		}
+		return false;
+	}
+
 	@Override
 	public void updateEntity() {
 		super.updateEntity();
+
 		if(worldObj.isRemote) {
 			return;
 		}
-		if(locomotive == null && !isInitialized && isBound) {
-			this.tryFindLocomotive(this.uuid);
-			if(locomotive == null) {
-				this.tryFindLocomotiveExpensively();
-				if(locomotive == null) {
-					Computronics.log.error(String.format(Locale.ENGLISH,
-						"Unable to find Electric Locomotive at (%s,%s,%s) bound to Locomotive Relay at (%s,%s,%s). It will not be bound anymore.",
-						locomotiveX, locomotiveY, locomotiveZ, xCoord, yCoord, zCoord));
-					if(uuid != null) {
-						Computronics.log.error("The Locomotive's UUID was: " + this.uuid);
-					}
-				}
-			}
-			isInitialized = true;
-		}
-		if(locomotive != null || !isBound) {
+
+		if(!isBound) {
 			return;
 		}
-		// Only check every second
-		if(worldObj.getTotalWorldTime() % 20 == 0) {
+
+		if(uuid == null) {
+			isBound = false;
+		}
+
+		if(uuid != null) {
 			this.tryFindLocomotive(this.uuid);
 		}
 	}
 
 	private void tryFindLocomotive(UUID uuid) {
-		if(locomotive == null) {
-			return;
-		}
 		if(uuid != null) {
-			EntityMinecart cart = CartTools.getLinkageManager(worldObj).getCartFromUUID(uuid);
-			if(cart != null && cart instanceof EntityLocomotiveElectric) {
-				this.setLocomotive((EntityLocomotiveElectric) cart);
-			}
-		}
-	}
-
-	private void tryFindLocomotiveExpensively() {
-		if(locomotive == null) {
-			return;
-		}
-		List locos = worldObj.getEntitiesWithinAABB(EntityLocomotiveElectric.class, AxisAlignedBB.getBoundingBox(locomotiveX - 0.5, locomotiveY - 0.5, locomotiveZ - 0.5,
-			locomotiveX + 0.5, locomotiveY + 0.5, locomotiveZ + 0.5));
-
-		EntityLocomotiveElectric checkLoco = null;
-		for(Object loco : locos) {
-			if(loco instanceof EntityLocomotiveElectric) {
-				if(checkLoco == null) {
-					checkLoco = (EntityLocomotiveElectric) loco;
-					continue;
+			EntityLocomotiveElectric cart = LocomotiveManager.instance().getCartFromUUID(uuid);
+			if(cart != null) {
+				EntityLocomotiveElectric oldLoco = getLocomotive();
+				if(oldLoco == null || cart != oldLoco) {
+					this.setLocomotive(cart);
 				}
-				EntityLocomotiveElectric newLoco = (EntityLocomotiveElectric) loco;
-				if(newLoco.getDistanceSq(locomotiveX, locomotiveY, locomotiveZ)
-					< checkLoco.getDistanceSq(locomotiveX, locomotiveY, locomotiveZ)) {
-					checkLoco = newLoco;
-				}
+			} else {
+				this.locomotive = null;
 			}
-		}
-		if(checkLoco != null) {
-			this.setLocomotive(checkLoco);
 		}
 	}
 
 	@Override
 	public void readFromNBT(NBTTagCompound nbt) {
 		super.readFromNBT(nbt);
-		if(this.locomotive == null) {
-			this.locomotiveX = nbt.getDouble("locomotiveX");
-			this.locomotiveY = nbt.getDouble("locomotiveY");
-			this.locomotiveZ = nbt.getDouble("locomotiveZ");
+		this.isBound = nbt.getBoolean("bound");
+		if(isBound) {
 			this.uuid = MiscTools.readUUID(nbt, "locomotive");
-			this.isBound = nbt.getBoolean("bound");
 		}
 	}
 
 	@Override
 	public void writeToNBT(NBTTagCompound nbt) {
 		super.writeToNBT(nbt);
-		if(this.locomotive != null) {
-			nbt.setDouble("locomotiveX", this.locomotive.posX);
-			nbt.setDouble("locomotiveY", this.locomotive.posY);
-			nbt.setDouble("locomotiveZ", this.locomotive.posZ);
-			MiscTools.writeUUID(nbt, "locomotive", this.locomotive.getPersistentID());
+		if(isBound && this.uuid != null) {
+			MiscTools.writeUUID(nbt, "locomotive", this.uuid);
 		}
 		nbt.setBoolean("bound", isBound);
 	}
@@ -150,9 +117,7 @@ public class TileLocomotiveRelay extends TileEntityPeripheralBase {
 	@Override
 	public void writeToRemoteNBT(NBTTagCompound nbt) {
 		super.writeToRemoteNBT(nbt);
-		if(this.locomotive != null) {
-			nbt.setBoolean("bound", isBound);
-		}
+		nbt.setBoolean("bound", isBound);
 	}
 
 	@Override
@@ -163,18 +128,42 @@ public class TileLocomotiveRelay extends TileEntityPeripheralBase {
 		}
 	}
 
-	private String cannotAccessLocomotive() {
-		if(this.locomotive == null) {
+	private String cannotAccessLocomotive(double amount) {
+		EntityLocomotiveElectric locomotive = getLocomotive();
+		if(!isBound) {
 			return "relay is not bound to a locomotive";
 		}
-		if(this.locomotive.dimension != this.worldObj.provider.dimensionId) {
+		if(locomotive == null) {
+			return "locomotive is currently not detectable";
+		}
+		if(locomotive.dimension != this.worldObj.provider.dimensionId) {
 			return "relay and locomotive are in different dimensions";
 		}
-		if(!(this.locomotive.getDistance(xCoord, yCoord, zCoord) <= Config.LOCOMOTIVE_RELAY_RANGE)) {
+		if(locomotive.getDistanceSq(xCoord, yCoord, zCoord) > Config.LOCOMOTIVE_RELAY_RANGE * Config.LOCOMOTIVE_RELAY_RANGE) {
 			return "locomotive is too far away";
 		}
-		if(this.locomotive.isSecure()) {
+		if(locomotive.isSecure()) {
 			return "locomotive is locked";
+		}
+		if(Config.LOCOMOTIVE_RELAY_CONSUME_CHARGE && (locomotive.getChargeHandler().getCharge() <= 0
+			|| locomotive.getChargeHandler().removeCharge(10 * amount) < 10 * amount)) {
+			return "locomotive out of energy";
+		}
+		if(Loader.isModLoaded(Mods.OpenComputers)) {
+			return cannotAccessLocomotive_OC(amount);
+		}
+		return null;
+	}
+
+	@Optional.Method(modid = Mods.OpenComputers)
+	private boolean tryConsumeEnergy(double energy) {
+		return node() instanceof Connector && ((Connector) node()).tryChangeBuffer(-energy);
+	}
+
+	@Optional.Method(modid = Mods.OpenComputers)
+	private String cannotAccessLocomotive_OC(double amount) {
+		if(!tryConsumeEnergy(Config.LOCOMOTIVE_RELAY_BASE_POWER * amount)) {
+			return "not enough energy";
 		}
 		return null;
 	}
@@ -195,47 +184,52 @@ public class TileLocomotiveRelay extends TileEntityPeripheralBase {
 	@Callback(doc = "function():string; gets the destination the locomotive is currently set to")
 	@Optional.Method(modid = Mods.OpenComputers)
 	public Object[] getDestination(Context context, Arguments args) {
-		if(cannotAccessLocomotive() != null) {
-			return new Object[] { null, cannotAccessLocomotive() };
+		String error = cannotAccessLocomotive(1.0);
+		if(error != null) {
+			return new Object[] { null, error };
 		}
-		return new Object[] { this.locomotive.getDestination() };
+		return new Object[] { getLocomotive().getDestination() };
 	}
 
 	@Callback(doc = "function(destination:string):boolean; Sets the locomotive's destination; there needs to be a golden ticket inside the locomotive")
 	@Optional.Method(modid = Mods.OpenComputers)
 	public Object[] setDestination(Context c, Arguments a) {
-		if(cannotAccessLocomotive() != null) {
-			return new Object[] { null, cannotAccessLocomotive() };
+		String error = cannotAccessLocomotive(3.0);
+		if(error != null) {
+			return new Object[] { null, error };
 		}
 		a.checkString(0);
-		return TileLocomotiveRelay.setDestination(this.locomotive, a.toArray());
+		return TileLocomotiveRelay.setDestination(getLocomotive(), a.toArray());
 	}
 
 	@Callback(doc = "function():number; gets the current charge of the locomotive")
 	@Optional.Method(modid = Mods.OpenComputers)
 	public Object[] getCharge(Context context, Arguments args) {
-		if(cannotAccessLocomotive() != null) {
-			return new Object[] { null, cannotAccessLocomotive() };
+		String error = cannotAccessLocomotive(1.0);
+		if(error != null) {
+			return new Object[] { null, error };
 		}
-		return new Object[] { this.locomotive.getChargeHandler().getCharge() };
+		return new Object[] { getLocomotive().getChargeHandler().getCharge() };
 	}
 
 	@Callback(doc = "function():string; returns the current mode of the locomotive; can be RUNNING, IDLE or SHUTDOWN")
 	@Optional.Method(modid = Mods.OpenComputers)
 	public Object[] getMode(Context context, Arguments args) {
-		if(cannotAccessLocomotive() != null) {
-			return new Object[] { null, cannotAccessLocomotive() };
+		String error = cannotAccessLocomotive(1.0);
+		if(error != null) {
+			return new Object[] { null, error };
 		}
-		return new Object[] { this.locomotive.getMode().toString() };
+		return new Object[] { getLocomotive().getMode().toString() };
 	}
 
 	@Callback(doc = "function():string; returns the current name of the locomotive")
 	@Optional.Method(modid = Mods.OpenComputers)
 	public Object[] getName(Context context, Arguments args) {
-		if(cannotAccessLocomotive() != null) {
-			return new Object[] { null, cannotAccessLocomotive() };
+		String error = cannotAccessLocomotive(1.0);
+		if(error != null) {
+			return new Object[] { null, error };
 		}
-		return new Object[] { this.locomotive.func_95999_t() != null ? this.locomotive.func_95999_t() : "" };
+		return new Object[] { getLocomotive().func_95999_t() != null ? getLocomotive().func_95999_t() : "" };
 	}
 
 	@Override
@@ -249,28 +243,29 @@ public class TileLocomotiveRelay extends TileEntityPeripheralBase {
 	public Object[] callMethod(IComputerAccess computer, ILuaContext context, int method, Object[] arguments)
 		throws LuaException, InterruptedException {
 		if(method < getMethodNames().length) {
-			if(cannotAccessLocomotive() != null) {
-				return new Object[] { null, cannotAccessLocomotive() };
+			String error = cannotAccessLocomotive(method == 1 ? 3.0 : 1.0);
+			if(error != null) {
+				return new Object[] { null, error };
 			}
 			switch(method) {
 				case 0: {
-					return new Object[] { this.locomotive.getDestination() };
+					return new Object[] { getLocomotive().getDestination() };
 				}
 				case 1: {
 					if(arguments.length < 1 || !(arguments[0] instanceof String)) {
 						throw new LuaException("first argument needs to be a string");
 					}
 
-					return TileLocomotiveRelay.setDestination(this.locomotive, arguments);
+					return TileLocomotiveRelay.setDestination(getLocomotive(), arguments);
 				}
 				case 2: {
-					return new Object[] { this.locomotive.getChargeHandler().getCharge() };
+					return new Object[] { getLocomotive().getChargeHandler().getCharge() };
 				}
 				case 3: {
-					return new Object[] { this.locomotive.getMode().toString() };
+					return new Object[] { getLocomotive().getMode().toString() };
 				}
 				case 4: {
-					return new Object[] { this.locomotive.func_95999_t() != null ? this.locomotive.func_95999_t() : "" };
+					return new Object[] { getLocomotive().func_95999_t() != null ? getLocomotive().func_95999_t() : "" };
 				}
 			}
 		}